third_party/blink/web_tests/virtual/single-import-map/wpt_internal/import-maps/parsing_parsing-invalid-json.json-expected.txt

This is a testharness.js-based test.
[FAIL] Invalid JSON
  assert_equals: expected "{\\"imports\\":{},\\"integrity\\":{},\\"scopes\\":{}}" but got "null"
Harness: the test ran to completion.
